+++ tuneapi/apis/model_mistral.py
@@ -201,6 +201,7 @@ class Mistral(tt.ModelInterface):
         max_threads: int = 10,
         retry: int = 3,
         pbar=True,
+        **kwargs,
     ):
         return distributed_chat(
             self,
@@ -209,4 +210,5 @@ class Mistral(tt.ModelInterface):
             max_threads=max_threads,
             retry=retry,
             pbar=pbar,
+            **kwargs,
         )