+++ tuneapi/apis/model_tune.py
@@ -226,6 +226,7 @@ class TuneModel(tt.ModelInterface):
         max_threads: int = 10,
         retry: int = 3,
         pbar=True,
+        **kwargs,
     ):
         return distributed_chat(
             self,
@@ -234,4 +235,5 @@ class TuneModel(tt.ModelInterface):
             max_threads=max_threads,
             retry=retry,
             pbar=pbar,
+            **kwargs,
         )